% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{addFit}
\alias{addFit}
\title{Add a fit to a spectrum object}
\usage{
addFit(spec, fit)
}
\arguments{
\item{spec}{A \code{spectrum} object}

\item{fit}{A \code{tetmerFit} object}
}
\value{
An updated \code{spectrum} object with the fit appended
}
\description{
Appends a \code{tetmerFit} object to the \code{fits} slot of a
\code{spectrum} object.
}
