% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{addvertlines}
\alias{addvertlines}
\title{Add vertical lines to k-mer spectrum plot}
\usage{
addvertlines(input, optimised = 0)
}
\arguments{
\item{input}{Input object from GUI}

\item{optimised}{The fit obtained from optim.}
}
\description{
Using \code{input}, this function plots dashed vertical lines at multiples of \code{n}
corresponding to the samples's ploidy.
}
\keyword{internal}
