% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{allowSegTet}
\alias{allowSegTet}
\title{Enable the segmental allotetraploid model}
\usage{
allowSegTet()
}
\value{
NULL, invisibly
}
\description{
Adds the segmental allotetraploid model (\code{"tse"}) to the list
of available models in the Tetmer UI. This model is hidden by default
as it is experimental. Call this function before launching the app
with \code{tetmer()}.
}
\examples{
\dontrun{
allowSegTet()
tetmer(E028)
}
}
