% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{asAutoFitConfig}
\alias{asAutoFitConfig}
\title{Normalize Shiny input or config into autofit config}
\usage{
asAutoFitConfig(x)
}
\arguments{
\item{x}{Input from GUI or normalized autofit config}
}
\value{
A named list representing autofit configuration
}
\description{
Normalize Shiny input or config into autofit config
}
\keyword{internal}
