% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{buildExpectedSpectrum}
\alias{buildExpectedSpectrum}
\title{Build expected spectrum data from model parameters}
\usage{
buildExpectedSpectrum(model, par, xrange, name, k)
}
\arguments{
\item{model}{Model code}

\item{par}{Named numeric vector of model parameters}

\item{xrange}{Numeric vector of length 2 giving multiplicity limits}

\item{name}{Spectrum name}

\item{k}{Numeric k-mer size}
}
\value{
A \code{spectrum} object
}
\description{
Build expected spectrum data from model parameters
}
\keyword{internal}
