% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{buildOptimisationSpec}
\alias{buildOptimisationSpec}
\title{Build optimisation bounds and starting values}
\usage{
buildOptimisationSpec(input)
}
\arguments{
\item{input}{Input from GUI or compatible list}
}
\value{
A list containing \code{lower}, \code{upper}, and \code{start}
}
\description{
Build optimisation bounds and starting values
}
\keyword{internal}
