% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{drawPeakMarkers}
\alias{drawPeakMarkers}
\title{Draw model peak markers onto the active plot}
\usage{
drawPeakMarkers(cov, nPeaks, ypos, col = "grey", cex = 4)
}
\arguments{
\item{cov}{Numeric, expected monoploid coverage}

\item{nPeaks}{Integer number of expected peaks}

\item{ypos}{Numeric y-position for the labels}

\item{col}{Character color used for lines and labels}

\item{cex}{Numeric text size for labels}
}
\description{
Draw model peak markers onto the active plot
}
\keyword{internal}
