% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{evalCurrentModel}
\alias{evalCurrentModel}
\title{Evaluate the current model over an x-range}
\usage{
evalCurrentModel(input, probs, factors, xmin, xmax, optimised = NULL)
}
\arguments{
\item{input}{Input from GUI or compatible list}

\item{probs}{Expression for k-mer spectrum peak shapes}

\item{factors}{Expression for k-mer spectrum peak factors}

\item{xmin}{Integer, lower x limit}

\item{xmax}{Integer, upper x limit}

\item{optimised}{Optimisation result; leave \code{NULL} for manual mode}
}
\value{
Numeric vector of expected k-mer counts
}
\description{
Evaluate the current model over an x-range
}
\keyword{internal}
