% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getFactors}
\alias{getFactors}
\title{Factors for k-mer spectrum peaks}
\usage{
getFactors(input)
}
\arguments{
\item{input}{Input from GUI (contains model etc.)}
}
\value{
An expression of the factors by which the k-mer spectrum peaks are multiplied.
}
\description{
Factors for k-mer spectrum peaks
}
\keyword{internal}
