% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getModelCode}
\alias{getModelCode}
\title{Extract a model code from a supported config object}
\usage{
getModelCode(x)
}
\arguments{
\item{x}{Model code or config-like object}
}
\value{
Character scalar
}
\description{
Extract a model code from a supported config object
}
\keyword{internal}
