% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getModelLabel}
\alias{getModelLabel}
\title{Get the display label for a model}
\usage{
getModelLabel(model)
}
\arguments{
\item{model}{Model code}
}
\value{
Character scalar
}
\description{
Get the display label for a model
}
\keyword{internal}
