% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getModelParameterNames}
\alias{getModelParameterNames}
\title{Get optimisation parameter names for a model}
\usage{
getModelParameterNames(model)
}
\arguments{
\item{model}{Model code}
}
\value{
Character vector
}
\description{
Get optimisation parameter names for a model
}
\keyword{internal}
