% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getModelParameters}
\alias{getModelParameters}
\title{Extract model parameters from manual or optimised inputs}
\usage{
getModelParameters(input, optimised = NULL)
}
\arguments{
\item{input}{Input from GUI or compatible list}

\item{optimised}{Optimisation result; leave \code{NULL} for manual mode}
}
\value{
Named list of numeric parameter values
}
\description{
Extract model parameters from manual or optimised inputs
}
\keyword{internal}
