% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getModelRangeInputs}
\alias{getModelRangeInputs}
\title{Get parameter range inputs for a model}
\usage{
getModelRangeInputs(input)
}
\arguments{
\item{input}{Input from GUI or compatible list}
}
\value{
Named list of parameter ranges
}
\description{
Get parameter range inputs for a model
}
\keyword{internal}
