% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getProbs}
\alias{getProbs}
\title{Raw k-mer spectrum peaks}
\usage{
getProbs(input)
}
\arguments{
\item{input}{Input from GUI (contains model etc.)}
}
\value{
An expression corresponding to the size and number of k-mer spectrum peaks
}
\description{
Raw k-mer spectrum peaks
}
\keyword{internal}
