% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getStartingVals}
\alias{getStartingVals}
\title{Starting values for optimisation}
\usage{
getStartingVals(input)
}
\arguments{
\item{input}{Input from GUI (contains model, parameter ranges, etc.)}
}
\value{
A named vector of starting vlues
}
\description{
Starting values for optimisation
}
\keyword{internal}
