% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{makeAutoFitConfig}
\alias{makeAutoFitConfig}
\title{Create a normalized autofit config}
\usage{
makeAutoFitConfig(
  model,
  kcov,
  vf,
  log10theta,
  log10Mbp,
  xrange,
  diverg = c(0.1, 100),
  pallo = c(0.01, 0.99)
)
}
\arguments{
\item{model}{Model code}

\item{kcov}{Numeric vector of length 2 for coverage bounds}

\item{vf}{Numeric vector of length 2 for variance factor bounds}

\item{log10theta}{Numeric vector of length 2 for log10 theta bounds}

\item{log10Mbp}{Numeric vector of length 2 for log10 haploid genome
size bounds in Mbp}

\item{xrange}{Numeric vector of length 2 for fitting x-range}

\item{diverg}{Numeric vector of length 2 for divergence bounds}

\item{pallo}{Numeric vector of length 2 for p-allo bounds}
}
\value{
A named list representing autofit configuration
}
\description{
Create a normalized autofit config
}
\keyword{internal}
