% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{makeFitRecord}
\alias{makeFitRecord}
\title{Create a fit record from the current app state}
\usage{
makeFitRecord(input, optimised, k)
}
\arguments{
\item{input}{Input from Shiny GUI}

\item{optimised}{Result from \code{doOptimisation}}

\item{k}{Numeric k-mer length}
}
\value{
A \code{tetmerFit} object
}
\description{
Constructs a \code{tetmerFit} object from the current Shiny input
and optimisation result.
}
