% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{modelUsesDivergence}
\alias{modelUsesDivergence}
\title{Determine whether a model includes divergence time}
\usage{
modelUsesDivergence(model)
}
\arguments{
\item{model}{Model code}
}
\value{
Logical scalar
}
\description{
Determine whether a model includes divergence time
}
\keyword{internal}
