% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{modelUsesPallo}
\alias{modelUsesPallo}
\title{Determine whether a model includes p-allo}
\usage{
modelUsesPallo(model)
}
\arguments{
\item{model}{Model code}
}
\value{
Logical scalar
}
\description{
Determine whether a model includes p-allo
}
\keyword{internal}
