% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{normaliseExpectedSpectrumPar}
\alias{normaliseExpectedSpectrumPar}
\title{Normalise expected-spectrum parameters}
\usage{
normaliseExpectedSpectrumPar(par, model)
}
\arguments{
\item{par}{Named numeric vector or coercible list of model parameters}

\item{model}{Model code}
}
\value{
Named numeric vector ordered for the given model
}
\description{
Normalise expected-spectrum parameters
}
\keyword{internal}
