% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{normaliseExpectedSpectrumXrange}
\alias{normaliseExpectedSpectrumXrange}
\title{Normalise expected-spectrum x-range}
\usage{
normaliseExpectedSpectrumXrange(xrange)
}
\arguments{
\item{xrange}{Numeric vector of length 2}
}
\value{
Integer vector of length 2
}
\description{
Normalise expected-spectrum x-range
}
\keyword{internal}
