% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{normaliseSliderRanges}
\alias{normaliseSliderRanges}
\title{Validate and normalise slider range overrides}
\usage{
normaliseSliderRanges(x)
}
\arguments{
\item{x}{A named list of slider range overrides}
}
\value{
A named list of numeric scalar slider range values
}
\description{
Validate and normalise slider range overrides
}
\keyword{internal}
