% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{plotSpecApp}
\alias{plotSpecApp}
\title{Plot annotated spectrum}
\usage{
plotSpecApp(input, spec)
}
\arguments{
\item{input}{Parameters from GUI}

\item{spec}{Spectrum object}
}
\description{
Wrapper around \code{plot.spectrum} adding some other bits.
}
\keyword{internal}
