% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{prepare.spectrum}
\alias{prepare.spectrum}
\title{Prepare spectrum for plotting and fitting}
\usage{
prepare.spectrum(spe)
}
\arguments{
\item{spe}{A \code{spectrum} object}
}
\value{
A \code{spectrum} object.
}
\description{
Prepare spectrum for plotting and fitting
}
\keyword{internal}
