% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{resolveTetmerAppResult}
\alias{resolveTetmerAppResult}
\title{Resolve tetmer app return value}
\usage{
resolveTetmerAppResult(result, originalSpec)
}
\arguments{
\item{result}{Value returned by \code{shiny::runApp}}

\item{originalSpec}{Original \code{spectrum} used to launch the app}
}
\value{
A \code{spectrum} object
}
\description{
Resolve tetmer app return value
}
\keyword{internal}
