% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{show,spectrum-method}
\alias{show,spectrum-method}
\title{Show method for spectrum objects}
\usage{
\S4method{show}{spectrum}(object)
}
\arguments{
\item{object}{A \code{spectrum} object}
}
\description{
Show method for spectrum objects
}
