% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{show,tetmerFit-method}
\alias{show,tetmerFit-method}
\title{Show method for tetmerFit objects}
\usage{
\S4method{show}{tetmerFit}(object)
}
\arguments{
\item{object}{A \code{tetmerFit} object}
}
\description{
Show method for tetmerFit objects
}
