% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{tet.server}
\alias{tet.server}
\title{Run the Tetmer app}
\usage{
tet.server(input, output)
}
\arguments{
\item{input}{Default argument fo ar \code{shiny} server function. Leave empty.}

\item{output}{Default argument fo ar \code{shiny} server function. Leave empty.}
}
\description{
Run the Tetmer app
}
