% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{tetServer}
\alias{tetServer}
\title{Run the Tetmer app server}
\usage{
tetServer(input, output, session)
}
\arguments{
\item{input}{Default argument for a \code{shiny} server function. Leave empty.}

\item{output}{Default argument for a \code{shiny} server function. Leave empty.}

\item{session}{Default argument for a \code{shiny} server function. Leave empty.}
}
\description{
Run the Tetmer app server
}
