% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E028}
\alias{E028}
\title{K-mer spectrum of \emph{Euphrasia arctica} individual E028}
\format{
A \code{spectrum} object containing a name and data frame:
\describe{
\item{name}{The name E028}
\item{data}{A dataframe with coulmns \describe{
\item{mult}{K-mer multiplicity}
\item{count}{The number of different k-mer at a given multiplicity}
}
}
}
}
\source{
Becher et al. (2020) \url{https://doi.org/10.1016/j.xplc.2020.100105}
}
\usage{
E028
}
\description{
A k-mer spectrum if allotetraploid \emph{E. arctica}.
}
\examples{
plot(E028, log="xy")
plot(E028, xlim=c(0, 200), ylim=c(0, 10000000))
}
\keyword{datasets}
