% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{getPeakCount}
\alias{getPeakCount}
\title{Get expected peak count for a model}
\usage{
getPeakCount(model)
}
\arguments{
\item{model}{Model code}
}
\value{
Integer number of expected peaks
}
\description{
Get expected peak count for a model
}
\keyword{internal}
