% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{makeMinFun}
\alias{makeMinFun}
\title{Generate a function to be minimised}
\usage{
makeMinFun(input, probs, factors)
}
\arguments{
\item{input}{Input from GUI (contains model, parameter ranges, etc.)}

\item{probs}{Expression for k-mer spectrum peak shapes}

\item{factors}{Expression for k-mer spectrum peak factors}
}
\value{
A function serving as input to \code{optim}
}
\description{
This is then minimised by \code{optim}.
}
\keyword{internal}
