% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{makeMinFun}
\alias{makeMinFun}
\title{Generate a function to be minimised}
\usage{
makeMinFun(input)
}
\arguments{
\item{input}{Input from GUI (contains model, parameter ranges, etc.)}
}
\value{
A function servoing as input to \code{optim}
}
\description{
This is then mimimised by \code{optim}.
}
\keyword{internal}
