% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{plot.spectrum}
\alias{plot.spectrum}
\title{Plot a k-mer spectrum}
\usage{
\method{plot}{spectrum}(x, main, xlab, ylab, ...)
}
\arguments{
\item{x}{An object of class spectrum}

\item{main}{(optional) A string passed to \code{plot()}}

\item{xlab}{(optional) A string passed to \code{plot()}}

\item{ylab}{(optional) A string passed to \code{plot()}}

\item{...}{other keyword aruments to be passed to \code{plot()}}
}
\description{
Uses the plot function to generate a scatter plot of a k-mer spectrum.
}
\details{
It may be useful to set \code{log="xy"} or to limit the plotting range using \code{xlims} or \code{ylims}.
}
\examples{
plot(E030, log="xy")
plot(E030, xlim=c(0,200), ylim=c(0,10000000))
}
