% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{pointsFit}
\alias{pointsFit}
\title{Plot the fit}
\usage{
pointsFit(input, optimised = 0)
}
\arguments{
\item{input}{From GUI}

\item{optimised}{From fit (only supplied in auto mode)}
}
\description{
Plot the fit
}
\keyword{internal}
