% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{read.spectrum}
\alias{read.spectrum}
\title{Read in a k-mer spectrum}
\usage{
read.spectrum(f, nam = "MySpectrum", k = 0, cropAt = 1000, no0 = FALSE, ...)
}
\arguments{
\item{f}{A string indicating the path to the spectrum file}

\item{nam}{(optional) A string indicating the name of the spectrum}

\item{k}{(optional) A numeric indicating the k-mer length}

\item{cropAt}{(optional) An integer specifying the maximum multiplicity
value to retain. K-mers with multiplicity above this value are
discarded. Defaults to 1000.}

\item{no0}{(optional) A logical value. If \code{FALSE} (default),
\code{prepare.spectrum} is run on the data to insert missing
multiplicity values. If \code{TRUE}, this step is skipped.}

\item{...}{keyword arguments to be passed to \code{read.table}}
}
\value{
A \code{spectrum} object
}
\description{
Uses \code{read.table()} to read in a k-mer spectrum.
}
\examples{
\dontrun{testdf <- data.frame(mult=1:100, count = 100:1)}
\dontrun{tf <- tempfile()}
\dontrun{write.table(testdf, tf)}
\dontrun{sp <- read.spectrum(tf, nam = "Test spectrum")}
\dontrun{unlink(tf)}
\dontrun{plot(sp)}
}
