% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{read.spectrum}
\alias{read.spectrum}
\title{Read in a k-mer spectrum}
\usage{
read.spectrum(f, nam = "spectrum", ...)
}
\arguments{
\item{f}{A string indicating the path to th spectrum file}

\item{nam}{(otional) A string indicaing the name of the spectrum}

\item{...}{keyword arguments to be passed to \code{read.table()}}
}
\value{
A two-element list comprising a names string and a two-column data frame of the
k-mer spectrum
}
\description{
Uses \code{read.table()} to read in a k-mer spectrum.
}
\examples{
testdf <- data.frame(mult=1:100, count = 100:1)
tf <- tempfile()
write.table(testdf, tf)
sp <- read.spectrum(tf, nam = "Test spectrum")
unlink(tf)
plot(sp)
}
