% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{read.spectrum}
\alias{read.spectrum}
\title{Read in a k-mer spectrum}
\usage{
read.spectrum(f, nam = "MySpectrum", k = 0, ...)
}
\arguments{
\item{f}{A string indicating the path to th spectrum file}

\item{nam}{(otional) A string indicating the name of the spectrum}

\item{k}{(otional) A numeric indicating the k-mer length}

\item{...}{keyword arguments to be passed to \code{read.table()}}
}
\value{
A two-element list comprising a names string and a two-column data frame of the
k-mer spectrum
}
\description{
Uses \code{read.table()} to read in a k-mer spectrum.
}
\examples{
\dontrun{testdf <- data.frame(mult=1:100, count = 100:1)}
\dontrun{tf <- tempfile()}
\dontrun{write.table(testdf, tf)}
\dontrun{sp <- read.spectrum(tf, nam = "Test spectrum")}
\dontrun{unlink(tf)}
\dontrun{plot(sp)}
}
