% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\docType{class}
\name{spectrum-class}
\alias{spectrum-class}
\title{A named k-mer spectrum class}
\description{
A named k-mer spectrum class
}
\section{Slots}{

\describe{
\item{\code{name}}{A string indicating the name of the spectrum (or sample)}

\item{\code{data}}{A \code{data.frame} with numeric cloumns mult and count.}
}}

