% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{tetmer}
\alias{tetmer}
\title{Run the interactive Tetmer app}
\usage{
tetmer(sp = E028)
}
\arguments{
\item{sp}{A \code{spectrum} object as generated by \code{read.spectrum}.}
}
\value{
A \code{spectrum} object. Pressing Done appends the current
manual fit or current autofit result to the spectrum's \code{fits}
slot. If the app is closed without pressing Done, the original input
spectrum is returned unchanged.
}
\description{
Run the interactive Tetmer app
}
\examples{
\dontrun{result <- tetmer(E028)}
\dontrun{result <- tetmer(E030)}
}
