% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetmer.R
\name{textOut}
\alias{textOut}
\title{Generate text for Tetmer window}
\usage{
textOut(input, optimised)
}
\arguments{
\item{input}{Input from the GUI (contains plotting range, model, etc.)}

\item{optimised}{Fitted values from \code{optim}.}
}
\value{
A string of the fitted parameters produced by \code{renderText()}
to be displayed in the Tetmer window.
}
\description{
Generate text for Tetmer window
}
\keyword{internal}
